Move filters and Swipes to separate component for better reuse on mobile
