+++ src/pages/FriendsPage.tsx
@@ -1,13 +1,12 @@
-import { useState, useRef } from 'react'
-import { Box, Link } from '@mui/material'
+import { useState } from 'react'
+import { Box } from '@mui/material'
 import { makeStyles } from 'tss-react/mui'
 import UserProfile from 'components/userProfile/UserProfile'
 import UserProfileButton from 'components/userProfile/UserProfileButton'
 import { UserProfileData } from 'types/UserProfileData'
 import Friends from 'components/tabsMessagesFriends/Friends'
 import { useNavigate } from 'react-router-dom'
-import Swipes from 'components/swipes/Swipes'
-import NoMoreMatchesDialog from 'pages/NoMoreMatchesDialog'
+import SwipesWithFilters from 'components/swipes/SwipesWithFilters'
 
 const FriendsPage = () => {
   const { classes } = useStyles()
@@ -15,14 +14,6 @@ const FriendsPage = () => {
   const [friendsData, setFriendsData] = useState<UserProfileData | null>(null)
 
   const navigate = useNavigate()
-
-  const FiltersDialogRef = useRef<{
-    handleOpenNoMoreMatchesDialog: () => void
-  }>(null)
-
-  const handleOpenFiltersDialog = () => {
-    FiltersDialogRef.current?.handleOpenNoMoreMatchesDialog()
-  }
 
   const selectFriend = (user: UserProfileData) => {
     setFriendsData(user)
@@ -41,14 +32,10 @@ const FriendsPage = () => {
           <UserProfileButton startChat={startChat} />
         </Box>
       ) : (
-        <Box>
-          <Link className={classes.filters} onClick={handleOpenFiltersDialog}>
-            filters
-          </Link>
-          <Swipes />
+        <Box sx={{ marginTop: '-71px' }}>
+          <SwipesWithFilters />
         </Box>
       )}
-      <NoMoreMatchesDialog ref={FiltersDialogRef} title="Filters" />
     </Box>
   )
 }
