+++ src/pages/MessagesPage.tsx
@@ -1,11 +1,10 @@
-import { useState, useRef } from 'react'
+import { useState } from 'react'
 import {
   Box,
   Typography,
   Button,
   TextareaAutosize,
   Avatar,
-  Link,
 } from '@mui/material'
 import { makeStyles } from 'tss-react/mui'
 import Messages from 'components/tabsMessagesFriends/Messages'
@@ -15,21 +14,12 @@ import { UserChatProfile } from 'types/UserProfileData'
 import StartChatting from 'components/chat/StartChatting'
 import DisplayingChat from 'components/chat/DisplayingChat'
 import messages from '../components/chat/chat.json'
-import Swipes from 'components/swipes/Swipes'
-import NoMoreMatchesDialog from 'pages/NoMoreMatchesDialog'
+import SwipesWithFilters from 'components/swipes/SwipesWithFilters'
 
 const MessagesPage = () => {
   const { classes } = useStyles()
   const [selectedChat, setSelectedChat] = useState<UserChatProfile | null>(null)
   const userId = '1'
-
-  const FiltersDialogRef = useRef<{
-    handleOpenNoMoreMatchesDialog: () => void
-  }>(null)
-
-  const handleOpenFiltersDialog = () => {
-    FiltersDialogRef.current?.handleOpenNoMoreMatchesDialog()
-  }
 
   const frienId = messages.participants.find((el) => el !== userId)
 
@@ -100,15 +90,11 @@ const MessagesPage = () => {
             </Box>
           </Box>
         ) : (
-          <Box>
-            <Link className={classes.filters} onClick={handleOpenFiltersDialog}>
-              filters
-            </Link>
-            <Swipes />
+          <Box sx={{ marginTop: '-71px' }}>
+            <SwipesWithFilters />
           </Box>
         )}
       </Box>
-      <NoMoreMatchesDialog ref={FiltersDialogRef} title="Filters" />
     </Box>
   )
 }
